Several updates
moved the tests to a separate test file in the tests directory.
renamed get_volume.py to volume.py.
made sure the method "get_volume" works for float and ndarrays by using e.g. np.atleast_1d(pos_x)
Initiated v_avg as an array
Stated the docstring, sthat mm is an instance of RTDC_DataSet
separate method a part of the volume calculation
